Share mock config helper for compaction tests
